requirements.txt
requests==2.32.3
tqdm==4.67.1
# paddlepaddle-gpu==3.0.0b2
# paddlex==3.0.0b2

### langchain
langchain
langchain-community
langchain-core==0.1.30
langchain-text-splitters==0.0.1
langsmith==0.1.23
openai
chromadb

## rag
nltk==3.8.1
sqlalchemy==2.0.28
bcembedding==0.1.3
ragas==0.1.7
streamlit==1.32.2
# requests==2.31.0
dataclass-wizard==0.22.3
pymilvus==2.3.7
python-docx==1.1.0
rapidocr-onnxruntime==1.3.15
PyMuPDF==1.24.5
python-multipart==0.0.9
unstructured[all-docs]
charset-normalizer==3.2.0
langfuse==2.31.0
wired-table-rec==0.0.3
chromadb==0.5.2
sentencepiece==0.2.0
langchain_openai

# 地址解析工具
addressparser

markitdown[all]

jionlp
scalar-fastapi